requirements.txt
Django==5.1.6
django-jazzmin==3.0.1
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
django-environ-0.12.0